+++ SuperMario/Brick.h
@@ -27,7 +27,7 @@ public:
 	~CBrick();
 	virtual void Render();
 	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
-	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* colliable_objects);
+	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* colliable_objects);
 
 	int GetBrickType() { return Btype; }
 	int GetItemRemaining() { return sl; };
