+++ SuperMario/Item.h
@@ -24,6 +24,6 @@ public:
 	int GetItemID() { return itemID; }
 
 	virtual void Render();
-	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects);
+	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* coObjects);
 	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
 };
