+++ SuperMario/Koopas.cpp
@@ -19,7 +19,7 @@ CKoopas::CKoopas(int x)
 }
 
 
-void CKoopas::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
+void CKoopas::Update(ULONGLONG dt, vector<LPGAMEOBJECT>* coObjects)
 {
 	CGameObject::Update(dt, coObjects);
