+++ SuperMario/Plant.h
@@ -22,7 +22,7 @@ public:
 	bool CheckPlayerInZone(D3DXVECTOR4 player);
 
 	virtual void GetBoundingBox(float& left, float& top, float& right, float& bottom);
-	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects, D3DXVECTOR4 player);
+	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* coObjects, D3DXVECTOR4 player);
 	virtual void Render();
 	virtual void SetState(int state);
 };
