+++ SuperMario/SuperMushroom.h
@@ -8,5 +8,5 @@ public:
 	CSuperMushroom(D3DXVECTOR2 position);
 	~CSuperMushroom();
 
-	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects);
+	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* coObjects);
 };
