+++ SuperMario/VenusFireTrap.h
@@ -43,7 +43,7 @@ public:
 
 	bool CheckDistanceNear(D3DXVECTOR4 player);
 	virtual void GetBoundingBox(float& left, float& top, float& right, float& bottom);
-	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects, D3DXVECTOR4 player);
+	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* coObjects, D3DXVECTOR4 player);
 	virtual void Render();
 	virtual void SetState(int state);
 };
